@@ -10,7 +10,7 @@
     "homepage": "https://flareapp.io/ignition",
     "license": "MIT",
     "require": {
-        "php": "^7.2.5|^8.0",
+        "php": "^7.4|^8.0",
         "ext-json": "*",
         "ext-mbstring": "*",
         "spatie/flare-client-php": "^1.6",
